move community_edge_betweenness example to tests
