+++ tests/CMakeLists.txt
@@ -633,6 +633,7 @@ add_legacy_tests(
   community_walktrap
   graphlets
   igraph_community_eb_get_merges
+  igraph_community_edge_betweenness
   igraph_community_fluid_communities
   igraph_community_infomap
   igraph_community_leading_eigenvector2
