+++ tests/unit/igraph_community_edge_betweenness.out
@@ -1,0 +1,4 @@
+15 1 7 45 52 31 23 16 24 25 28 44 68 27 4 5 3 8 76 75 70 57 58 26 9 67 66 10 33 46 47 48 49 63 69 50 51 12 20 53 54 13 21 35 38 55 56 14 22 29 42 43 41 73 74 6 18 32 0 2 11 17 19 30 34 36 37 39 40 59 60 61 62 64 65 71 72 77 
+71.39 66.90 77.32 82.00 123.23 100.21 143.63 109.25 107.67 142.75 285.00 16.83 18.18 18.00 15.33 25.33 25.00 50.00 14.50 22.37 25.62 29.65 40.67 72.00 9.00 9.00 11.00 5.50 8.00 5.00 10.00 4.50 9.00 4.50 9.00 4.00 8.00 3.50 7.00 3.50 7.00 3.00 6.00 3.00 6.00 3.00 6.00 2.50 5.00 2.00 2.00 3.50 5.00 2.00 4.00 1.33 2.00 4.00 1.00 1.50 3.00 1.00 2.00 1.00 1.00 1.00 1.00 2.00 1.00 1.00 1.50 3.00 1.00 2.00 1.00 1.00 2.00 1.00 
+71.39 66.90 77.32 82.00 123.23 100.21 143.63 109.25 107.67 142.75 285.00 16.83 18.18 18.00 15.33 25.33 25.00 50.00 14.50 22.37 25.62 29.65 40.67 72.00 9.00 9.00 11.00 5.50 8.00 5.00 10.00 4.50 9.00 4.50 9.00 4.00 8.00 3.50 7.00 3.50 7.00 3.00 6.00 3.00 6.00 3.00 6.00 2.50 5.00 2.00 2.00 3.50 5.00 2.00 4.00 1.33 2.00 4.00 1.00 1.50 3.00 1.00 2.00 1.00 1.00 1.00 1.00 2.00 1.00 1.00 1.50 3.00 1.00 2.00 1.00 1.00 2.00 1.00 
+No crash
